Fix Typo Everywhere u32 vs u16.
Whoops. Made priority u32 instead of u16 then propagated throughout the
codebase. Fixed.

+++ src/bulk_enqueue.rs
@@ -154,7 +154,7 @@ mod tests {
     impl JobKind for SendEmail {
         const NAME: &'static str = "send_email";
         const QUEUE: &'static str = "emails";
-        const PRIORITY: Option<u32> = Some(50);
+        const PRIORITY: Option<u16> = Some(50);
     }
 
     #[derive(Serialize, Deserialize)]
